+++ addon/components/flexberry-simpledatetime.js
@@ -233,6 +233,14 @@ export default FlexberryBaseComponent.extend({
   classNames: ['flexberry-simpledatetime'],
 
   /**
+    Array CSS class names for scroll.
+
+    @property scrollClassNames
+    @type Array
+  */
+  scrollSelectors: ['.full.height'],
+
+  /**
     If true, then onClick calling flatpickr.open().
 
     @property canClick
@@ -264,6 +272,7 @@ export default FlexberryBaseComponent.extend({
     this._super(...arguments);
     if (!(this.get('useBrowserInput') && this.get('currentTypeSupported'))) {
       this._flatpickrCreate();
+      Ember.$(this.scrollSelectors.join()).scroll(() => this.get('_flatpickr').close());
     }
   },
 
@@ -276,6 +285,7 @@ export default FlexberryBaseComponent.extend({
   willDestroyElement() {
     this._super(...arguments);
     this._flatpickrDestroy();
+    Ember.$(this.scrollSelectors.join()).unbind('scroll');
   },
 
   /**